Split test factories and integrate memex factories
Split the test factories module up into a package where each test
factory (or group of closely-related test factories) is in its own
module. This also moves the previously-memex-provided factories
(Annotation, Document, etc.) into tests.common.
