+++ tests/common/factories/feature.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+
+from h import models
+
+from .base import ModelFactory
+
+
+class Feature(ModelFactory):
+
+    class Meta:
+        model = models.Feature
