+++ tests/common/factories/feature_cohort.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+
+from h import models
+
+from .base import ModelFactory
+
+
+class FeatureCohort(ModelFactory):
+
+    class Meta:
+        model = models.FeatureCohort
